move CreateBranchOptions
