+++ GitRepository.go
@@ -23,7 +23,7 @@ type GitRepository interface {
 	CloneRepositoryByPathOrUrl(pathOrUrl string, verbose bool) (err error)
 	Commit(commitOptions *GitCommitOptions) (createdCommit *GitCommit, err error)
 	CommitHasParentCommitByCommitHash(hash string) (hasParentCommit bool, err error)
-	CreateBranch(createOptions *CreateBranchOptions) (err error)
+	CreateBranch(createOptions *parameteroptions.CreateBranchOptions) (err error)
 	Create(verbose bool) (err error)
 	CreateFileInDirectory(verbose bool, filePath ...string) (createdFile File, err error)
 	CreateSubDirectory(subDirectoryName string, verbose bool) (createdSubDirectory Directory, err error)
@@ -77,7 +77,7 @@ type GitRepository interface {
 	MustCommit(commitOptions *GitCommitOptions) (createdCommit *GitCommit)
 	MustCommitHasParentCommitByCommitHash(hash string) (hasParentCommit bool)
 	MustCreate(verbose bool)
-	MustCreateBranch(createOptions *CreateBranchOptions)
+	MustCreateBranch(createOptions *parameteroptions.CreateBranchOptions)
 	MustCreateFileInDirectory(verbose bool, filePath ...string) (createdFile File)
 	MustCreateSubDirectory(subDirectoryName string, verbose bool) (createdSubDirectory Directory)
 	MustCreateTag(createOptions *GitRepositoryCreateTagOptions) (createdTag GitTag)
